pyproject.toml
[project]
name = "yc-atlas-api"
version = "0.1.0"
description = ""
authors = [
    {name = "piyushk6626",email = "kulkarnipiyush462@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pymongo (>=4.11.3,<5.0.0)",
    "bson (>=0.5.10,<0.6.0)",
    "openai (>=1.68.2,<2.0.0)",
    "pinecone (>=6.0.2,<7.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
